// src/boilerplate/cmsApis/models/campaign.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Configuration Management
 * Collection of all Configuration Management Requests
 *
 * The version of the OpenAPI document: v6.7.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface Campaign {
    /**
     * Unique Id of the Campaign
     */
    'campaignContextId'?: number;
    /**
     * Name of the Campaign
     */
    'campaignContextName'?: string;
    /**
     * Type of the Campaign
     */
    'campaignContextType'?: CampaignCampaignContextTypeEnum;
    /**
     * Unique Id of the assigned Contact Center
     */
    'contactCenterId'?: number;
    /**
     * Unique Id of the assigned Process
     */
    'processId'?: number;
    /**
     * Campaign Version
     * @deprecated
     */
    'campaignContextVersion'?: string;
    /**
     * Campaign Description
     */
    'description'?: string;
    /**
     * Entity date added
     */
    'dateAdded'?: string;
    /**
     * Entity date modified
     */
    'dateModified'?: string;
    /**
     * Patch version of the entity
     */
    'ref'?: number;
    /**
     * Schema version of the entity
     */
    'version'?: number;
}

export const CampaignCampaignContextTypeEnum = {
    InteractionCampaign: 'Interaction Campaign',
    ChatCampaign: 'Chat Campaign',
    InteractiveVoiceApplication: 'Interactive Voice Application',
} as const;

export type CampaignCampaignContextTypeEnum = typeof CampaignCampaignContextTypeEnum[keyof typeof CampaignCampaignContextTypeEnum];